Create TexasHoldem.sas
Texas Hold’em is played with a standard deck of 52 cards. The object of the game is to make the best possible five-card poker hand using any combination of the two card’s in a player’s hand (known as the “hole cards”) and the five community cards that the dealer places in the middle of the table, which can be used by all players. The standard hand rankings are used, as follows (from best to worst):

Straight Flush: Any five cards of the same suit and consecutive rank; the best of these hands, AKQJT of a single suit, is known as a Royal Flush. For these and regular straights, aces may be played as high or low cards.
Four of a Kind: Any four cards of the same rank.
Full House: Any three cards from a single rank combined with a pair from a different rank (i.e., TTT55).
Flush: Any five cards of the same suit.
Straight: Any five cards of consecutive ranks (i.e., 34567).
Three of a Kind: Any three cards of the same rank.
Two Pair: Any two pairs of cards from the same ranks (ie, 5599A).
One Pair: Any two cards of the same rank.
High Card: Hands that do not fit any of the above categories are ranked based on the highest card in their hand (aces are high), then by the second highest card, and so on.